restructured controller folders
