+++ src/controllers/decode/decode.ts
