+++ src/controllers/encode/encode.ts
@@ -5,7 +5,7 @@ export const encodeUrl = ((req: Request, res: Response) => {
     const url: string = req.body.url
 
     if (!validUrl.isUri(url)){
-        // res.setStatus(400)
+        res.sendStatus(400)
     } 
 
     res.send("encodedUrl Controller")
